+++ CHANGELOG.md
@@ -7,6 +7,7 @@ v1.1.0 [2018-02-24]
 - Added tests.
 - Fixed table options in a migration.
 - Small refactoring.
+- Fixed condtion of `skipAttributes`.
 
 v1.0.0 [2016-01-04]
 -------------------